backend/requirements.txt
Django>=5.0,<6.0
celery>=5.3
deepagents>=0.2.8
django-cors-headers>=4.3
django-environ>=0.11
django-redis>=5.4
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
kubernetes>=28.1.0
langchain>=1.1.0,<1.2
langchain-community>=0.3.0
langchain-openai>=1.1.0,<1.2
langchain-ollama>=0.1.0
langgraph-checkpoint-postgres==3.0.1
opentelemetry-exporter-otlp>=1.24
opentelemetry-sdk>=1.24
playwright>=1.51.0
pydantic>=2.6
psycopg[binary]>=3.1
redis>=5.0
requests>=2.32
structlog>=24.1
tavily-python>=0.5.0
gunicorn>=21.2
whitenoise>=6.6
boto3>=1.34
uvicorn[standard]>=0.29
websockets>=12.0
